monitor/requirements.txt
tkinter
mysql-connector-python
requests
datetime
threading
pyserial
ctypes
